Replaced sklearn scalers by a single ScalingTransformer. Moved all pytest fixtures to conftest.py and changed their scopes to "session".
