+++ ml_project/README.md
@@ -1,11 +1,14 @@
 Игрушечная задачка предсказания наличия заболевания используя этот [датасет](https://www.kaggle.com/ronitf/heart-disease-uci)
 
-* для установки из ml_project ```pip install .```
+Все команды запускаются из ml_project:
+
+* для установки ```pip install .```
 
 * чтобы запустить ```python ml_pipeline/train_pipeline.py configs/train_config.yaml```
 
-* прогнать тесты ```python -m pytest ml_project/tests/```(тесты пока не работают)
+* прогнать тесты ```pytest tests```
 
-Тренировка и предсказание происходят в ```train_pipeline.py```. Сначала трансформером подготавливаются фичи: замещение нанов, one-hot для категориальных, скейлер для численных фич, если скейлер указан в конфиге. Обучить можно либо логрегрессию, либо RandomForest. Я использую два разных конфига для обучения и в тестах. Один с лесом и без скейлера, другой с логрегрессией и скейлером. Предсказание predict принимает на вход обученную модель и зафитированные трансформером тестовые фичи. В валидации считается roc-auc, f1 и recall.
+Тренировка и предсказание происходят в ```train_pipeline.py```. Сначала трансформером подготавливаются фичи: замещение нанов, one-hot для категориальных, скейлер для численных фич, если указан в конфиге. Если скейлер не null, то применяется кастомный трансформер, который отрезает outliers, подсчитывает среднее и стандартное отклонение
+и рескейлит данные. Обучить можно либо логрегрессию, либо RandomForest. Я использую два разных конфига для обучения и в тестах. Один с лесом и без скейлера, другой с логрегрессией и скейлером. Предсказание predict принимает на вход обученную модель и зафитированные трансформером тестовые фичи. В валидации считается roc-auc, f1 и recall.
 
 В notebooks есть небольшой файл с попарными распределениями фич, трансформерами, парой моделек и генерацией фейковых данных.
