+++ ml_project/ml_pipeline/features/__init__.py
@@ -1,8 +1,10 @@
 from .build_features import (
+    ScalingTransformer,
     build_transformer,
     transform_features,
     get_target)
 
-__all__ = ["build_transformer",
+__all__ = ["ScalingTransformer",
+           "build_transformer",
            "transform_features",
            "get_target"]
