+++ ml_project/tests/features/__init__.py
@@ -1,3 +1,3 @@
-from .test_process_features import feature_params, preprocessing_params
+from .test_process_features import CATEGORICAL, NUMERICAL, SCALER, TARGET
 
-__all__ = ["feature_params", "preprocessing_params"]
+__all__ = ["CATEGORICAL", "NUMERICAL", "SCALER", "TARGET"]
