+++ ml_project/tests/features/test_process_features.py
@@ -1,62 +1,54 @@
 import numpy as np
-import pytest
-from sklearn.preprocessing import StandardScaler, MinMaxScaler
-from ml_pipeline.params import FeatureParams, PreprocessingParams
-from ml_pipeline.features import (build_transformer,
-                                 transform_features,
-                                 get_target)
-from tests.data import fake_data, data
+from ml_pipeline.features import (ScalingTransformer,
+                                  build_transformer,
+                                  transform_features,
+                                  get_target)
 
 
 CATEGORICAL = ["sex", "cp", "fbs", "restecg", "exang", "slope", "ca", "thal"]
 NUMERICAL = ["age", "trestbps", "chol", "thalach", "oldpeak"]
 TARGET = "target"
-SCALER = "StandardScaler"
+SCALER = "true"
 
 
+def test_can_scale_features(fake_data):
+    transformer = ScalingTransformer()
+    num_data = fake_data[NUMERICAL].dropna()
+    transformer.fit(num_data)
+    transformed_num_features = transformer.transform(num_data)
 
-@pytest.fixture
-def feature_params() -> FeatureParams:
-    params = FeatureParams(
-        categorical_features=CATEGORICAL,
-        numerical_features=NUMERICAL,
-        target_col=TARGET,
-    )
-    return params
+    assert transformed_num_features.shape[1] == len(NUMERICAL), \
+        f"shape of scaled features is {transformed_num_features.shape[1]}, " \
+        f"expected {len(NUMERICAL)}"
 
+    assert np.allclose(transformed_num_features.mean(), 0, atol=1), \
+        f"mean of scaled features not 0: {transformed_num_features.mean()}"
+    assert np.allclose(transformed_num_features.std(), 1, atol=1), \
+        f"std of scaled features is not 1: {transformed_num_features.std()}"
 
-@pytest.fixture
-def preprocessing_params() -> PreprocessingParams:
-    params = PreprocessingParams(
-        scaler=SCALER
-    )
-    return params
+    upper_percentile = np.percentile(num_data, 95, axis=0)
+    lower_percentile = np.percentile(num_data, 5, axis=0)
+    clipped_num_features = num_data[(num_data < upper_percentile) & (num_data > lower_percentile)]
+    expected_num_features = (num_data - np.mean(clipped_num_features)) / np.std(clipped_num_features)
+
+    assert transformed_num_features.equals(expected_num_features), \
+        "scaled features are different from expected"
 
 
 def test_can_transform_features(fake_data,
-                            feature_params):
-    for scaler in ["StandardScaler", "MinMaxScaler"]:
-        prcs_params = PreprocessingParams(scaler=scaler)
-        transformer = build_transformer(feature_params, prcs_params)
-        transformer.fit(fake_data)
-        features = transform_features(transformer, fake_data)
+                                feature_params,
+                                preprocessing_params):
 
-        numerical_features = features.iloc[:, : len(NUMERICAL)]
-        if scaler == "StandardScaler":
-            assert np.allclose(numerical_features.mean(axis=0), 0, atol=1e-1), \
-                "mean of scaled features not 0"
-            assert np.allclose(numerical_features.std(axis=0), 1, atol=1e-1), \
-                "std of scaled features not 1"
-        elif scaler == "MinMaxScaler":
-            assert np.all(abs(numerical_features) <= 1), \
-                "scaled features are outside of [-1, 1] range"
+    transformer = build_transformer(feature_params, preprocessing_params)
+    transformer.fit(fake_data)
+    features = transform_features(transformer, fake_data)
 
-        assert features.isna().sum().sum() == 0, \
-            "NaNs are present after transform"
-        assert features.shape[0] == fake_data.shape[0], \
-            "features must contain the same number of rows as original data"
-        assert features.shape[1] > fake_data.shape[1], \
-            "categorical features were not one hot encoded"
+    assert features.isna().sum().sum() == 0, \
+        "NaNs are present after transform"
+    assert features.shape[0] == fake_data.shape[0], \
+        "features must contain the same number of rows as original data"
+    assert features.shape[1] > fake_data.shape[1], \
+        "categorical features were not one hot encoded"
 
 
 def test_can_get_target(fake_data, feature_params):
