+++ ml_project/tests/models/test_train_model.py
@@ -1,35 +1,15 @@
-import pytest
 import pandas as pd
 from typing import Tuple
 from sklearn.ensemble import RandomForestClassifier
 import pickle
 from py._path.local import LocalPath
 
-from tests.data import fake_data, data
-from tests.features import feature_params, preprocessing_params
-from ml_pipeline.features import (build_transformer,
-                                 transform_features,
-                                 get_target)
 from ml_pipeline.params import TrainingParams
-
 from ml_pipeline.models.utils import (
     train_model,
     predict,
     save_model
 )
-
-
-@pytest.fixture
-def features_and_target(fake_data,
-                        feature_params,
-                        preprocessing_params):
-    transformer = build_transformer(feature_params,
-                                    preprocessing_params)
-    transformer.fit(fake_data)
-    features = transform_features(transformer, fake_data)
-    target = get_target(fake_data, feature_params)
-
-    return features, target
 
 
 def test_can_train_model(features_and_target: Tuple[pd.DataFrame, pd.Series]):
